% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvmmat_est.R
\name{glmm_est}
\alias{glmm_est}
\title{NULL model estimation for GMMAT and RGMMAT}
\usage{
glmm_est(
  y.long,
  time,
  y.cov,
  phe.model = phe.model,
  maxiter = 50,
  tol = 10^(-6)
)
}
\arguments{
\item{y.long}{Long-formatted phenotype vector}

\item{time}{Time covarites matched with phenotype vector}

\item{y.cov}{Covariate matrix denoting the covariate variables measured at each time}

\item{phe.model}{String, the phenotype model, two optional values: 'logistic', 'Gaussian'}

\item{maxiter}{Numeric, the maximum count for the iterative estimation in when using mixture correlation structure}

\item{tol}{Numeric, tolerance for the iterative estimation in when using mixture correlation structure}
}
\value{
This function returns a list object with model parameters and residuals of the NULL generalized linear mixed effect model
}
\description{
This function estimate the parameters and residuals for the NULL generalized linear mixed effect model
}
