% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvmmat_est.R
\name{rvmmat_est}
\alias{rvmmat_est}
\title{NULL model estimation}
\usage{
rvmmat_est(
  y.long,
  time,
  y.cov,
  phe.model = phe.model,
  maxiter = 50,
  tol = 10^(-6),
  Hd = 2,
  VCcorrection = FALSE
)
}
\arguments{
\item{y.long}{Long-formatted phenotype vector}

\item{time}{Time covarites matched with phenotype vector}

\item{y.cov}{Covariate matrix denoting the covariate variables measured at each time}

\item{phe.model}{String, the phenotype model, two optional values: 'logistic', 'liability'}

\item{maxiter}{Numeric, the maximum count for the iterative estimation in when using mixture correlation structure}

\item{tol}{Numeric, tolerance for the iterative estimation in when using mixture correlation structure}

\item{Hd}{Numeric, degree of derivative of smooth function which lead to smooth spline of order 2Hd}

\item{VCcorrection}{Logical variable, indicating whether apply correction for variance components and natural parameters}
}
\value{
This function returns a list object with model parameters and residuals of the NULL model
}
\description{
This function estimate the parameters and residuals for the NULL model in RVMMAT
}
